new migrations and reorganized old

+++ database/migrations/2021_01_16_201604_create_klikbud_main_slider_table.php
@@ -17,8 +17,8 @@ class CreateKlikbudMainSliderTable extends Migration
             $table->id();
             $table->tinyInteger('status_to_main_page_id')->unsigned()->nullable(); //  1-active, 2-hidden(but show), 3-disable
             $table->integer('slider_number_show')->nullable(); //1,2,3; 0 - default
-            $table->integer('image_id')->unsigned()->nullable(); //id
-            $table->integer('user_id')->unsigned()->nullable();
+            $table->bigInteger('image_id')->unsigned()->nullable(); //id
+            $table->bigInteger('user_id')->unsigned()->nullable();
             $table->tinyInteger('moderated_id')->unsigned()->nullable();
             $table->json('alt')->nullable();
             $table->json('textYellow')->nullable();
