+++ database/migrations/2021_01_16_202732_create_klikbud_gallery_table.php
@@ -15,11 +15,11 @@ class CreateKlikbudGalleryTable extends Migration
     {
         Schema::create('klikbud_gallery', function (Blueprint $table) {
             $table->id();
-            $table->integer('object_id')->unsigned()->nullable();
+            $table->bigInteger('object_id')->unsigned()->nullable();
             $table->integer('category_id')->unsigned()->nullable();
             $table->tinyInteger('status_gallery_id')->unsigned()->nullable(); // Status to view gallery
             $table->tinyInteger('status_to_main_page_id')->unsigned()->nullable(); // In main page 1
-            $table->integer('image_id')->unsigned()->nullable();
+            $table->bigInteger('image_id')->unsigned()->nullable();
             $table->json('old_images_id')->nullable();
             $table->json('slug')->nullable();
             $table->json('title')->nullable();
