+++ database/migrations/2021_01_25_180618_create_klikbud_opinion_table.php
@@ -20,9 +20,9 @@ class CreateKlikbudOpinionTable extends Migration
             $table->tinyInteger('stars')->unsigned()->nullable();
             $table->tinyInteger('status_to_main_page_id')->unsigned()->default(2); // 1 - Active in main page 2 - Hidden
             $table->tinyInteger('is_new_id')->unsigned()->default(false); // New where client send new opinion in site
-            $table->integer('reader_id')->nullable(); //If status is new id
+            $table->bigInteger('reader_id')->nullable(); //If status is new id
             $table->integer('portal_opinion_id')->unsigned()->default(NULL);
-            $table->integer('user_id')->unsigned()->nullable();
+            $table->bigInteger('user_id')->unsigned()->nullable();
             $table->integer('moderated_id')->default(3);
             $table->mediumText('opinion')->nullable();
             $table->dateTime('date_add')->nullable();
